Pop up window behavior
New game windows now use a normal NSWindow title bar. Im happy to
switch this back, but I had many crashes re-sizing, and could not
properly position the window. New window size is now large enough to
contain the hover utility bar in the window without it overflowing, and
windows are now centered when opened, so they are in a nice location
for the user.