Try to optimize OEScrollView background rendering.
This change set replaces -drawRect: calls with creating a layer
responsible for storing and resizing the background image. Another layer
is responsible for displaying the noise. The noise image is converted into
a pattern, the pattern is then converted into a color. That color is used
as a background color to the 'noiseLayer'. These minor modifications seems
to improve CPU usage during live resizing.