Clean up warning in NSWindow+OECustomWindow
ARC doesn't like -performSelector:withObject: because it doesn't know
the memory semantics of the result value. Switch this code over to using
objc_msgSend() instead. While we're at it, make the called method take
an NSRect instead of an NSValue, since that was a hack for use with
-performSelector:withObject:. However, we're not actually implementing
this method anywhere in the project, so we don't need to update any
other methods.