Turn off ARC for NSColor+OEAdditions.m
NSColor+OEAdditions.m defines an accessor on NSColor for -CGColor. This
method is designed to return an autoreleased CGColorRef calculated from
the NSColor. Under ARC, there is no supported mechanism to return an
autoreleased CF/CG object from a method, and the fact that this was
working was undocumented behavior of the compiler that may change in
future versions. As such, we need to turn off ARC for this file so we
can explicitly call -autorelease on the object.