Update UI when there are changes in database.
Monitor the main thread's MOC for added, removed, and changed 'Game'
entities. If games are added or removed, then reload the views. If the games'
properties are simply changed then just update the visible games.

-reloadData is an expensive operation as it has to calculate all of the
internal variables.  Each view (gridView, tableView, and coverFlowView) have
their own methods for reloading only the data of the visible cells. This is a
less intensive operation (compared to -reloadData) as the majority of the
cached values will remain.