Remove visual bug when importing roms.
When a rom is imported and only the visible cells need to be loaded, sometimes
there will be a ghost image of a cell's text. This change will attempt to
retrieve the visible cell that needs to be updated, attempts to dequeue a
reusable cell, and as a last resort creates a new cell.

This eliminates redundant calls while trying to reload a cell's properties.