Partial reversion of the runloop stuff
It seems that the runloop occasionally misses the timer fire (if anything else
slows it down) which is enough to cause audio buffer underruns.

Note that we still are using timing based on video rather than audio so it's
still a coincidence that audio doesn't skip.

Also init audio slightly later so the starting buffer will have some data.