Easier way to calculate the selection box.
Retrieve the min of the initial and current points, get the max of the two, and
build a rectangle that covers those two points. The min points are restricted by
the minimum boundary {0,1} and the max points are restricted by the maximum
boundary {CGGetMaxX(bounds), CGGetMaxY(bounds)}.  The origin of the selection box
is the min points while the size is max points subtracted by the min points.