Add "aspectSize" to gameCore. This is the user interface aspect ratio of the game core. OEGameCore framework defaults to 4:3, but cores can override and even change per frame if they need to.
OEGameView now never deals directly with the buffer size of the game core. We only see the "screenSize" which is the internal dimensions of the incoming texture from our helper app, and the aspect, to consistently display the user facing screen of the emulator consistently, even though the emulator may change internal sizes.

Update the helper and extra bullshit to ensure this fuck-wittery actually works. Which it seems like it does. Holy fucking shit.