Teach OEMenu to look at modifiers when it opens
OEMenu only looks at key modifiers when it receives the -flagsChanged:
event. This means that it will never show alternate menuitems unless a
modifier that it's interested in changes, even if one or more
interesting modifiers were already active when the menu first opened.

This exhibits today as saves cannot be deleted unless Option is pressed
while the menu is already visible.

Fixes #362