Transparently decompress gzipped ROMs at runtime
Teach OpenEmuHelperApp to detect when the ROM it wants to launch is
compressed in a fashion libz can handle (e.g. gzip). When this is
detected, decompress the file into TMPDIR and launch from there instead.

It should be pretty easy to add other decompression types, such as
bzip2, if needed.

Fixes #368