IOHIDEvent.m: triggerEventWithPadNumber: use proper maximum.
If this is set to value, as it is before this change, then value often
(always?) is initialized to 0, and then when the trigger is actually
used, there is an assertion failure b/c the value (positive number) is
beyond the maximum (0).

This was occurring when using a Wii Remote with Classic Controller
Pro.