Support for new all new Wii Remotes with MotionPlus inside.
There are two types of Wii Remote at the time of writing: The
traditional ones and a newer variety that is more restrictive in how
one interacts with it. To support both, a few changes are required:

Discovery:

Newer Wii Remotes have a different minor device class than the older
ones, so a scan that simply looks for a particular major and minor
device class will miss at least one type of Wii Remote. To address
this, we scan for all bluetooth devices for 3 seconds, then check each
for a common prefix shared by the names of both types of Wii Remote
("Nintendo RVL-CNT-01" and "Nintendo RVL-CNT-01-TR").

Device Communication:

Historically, open-source Wii code has treated the Wii Remote like a
standard HID device and sent messages over the control channel with a
0x52 byte prefix, but unfortunately the newer Wii Remotes don't
receive messages sent like this. All Wii Remotes will accept messages
sent on the interrupt channel with a byte prefix of 0xa2, though, so
we change the code to use that.

To test this change, I paired 3 Wii Remotes simultaneously: One
Nintendo RVL-CNT-01 and two Nintendo RVL-CNT-01-TR. The -TR remotes
need to be disoverable by using the wrist strap to push the red button
on the back, while the non-TR one was made discoverable by holding
down buttons 1+2.