Move event handling from the UI process to the emulator process. Before this change, we needed 3 objects to be exchanged between the UI and emulator processes to have all the information. The last object was a proxy to the game core to receive the inputs from the UI process. This object is now removed and the device events are now handled on the emulator process side. This is achieved by moving the system responder to the emulator process which means it also needs to be set up on that side. To do so we add two OEGameCoreHelper methods that allows us to send the binding translations over to the emulator process.
This also means we move the so-called "global keys" (quick save, quick load, reset, pause, mute, etc.) management to the emulator process. Since the actual handle code is still on the UI process side we add these methods to the OEGameCoreOwner (formerly OEGameCoreDisplayHelper) protocol. OEGameDocument serves as the game core owner and handles all these new methods, it also sends the screen updates to its OEGameViewController and OEGameView. OpenEmuHelperApp now allocates both the core plug-in and system plug-in and their corresponding controllers to retrieve the events and pass them to the game core.

Finally, we also add a way to block events from being distributed to the game cores if the emulator window is not the frontmost one. Currently, this is handled using booleans but we could improve the whole architecture and allow different devices to be used on separate emulators running simultaneously by telling each emulator which devices they can use. Currently, this is only to prevent managing keyboard events if OpenEmu is not the frontmost app and preventing the emulators from handling any events while the preference window is shown.