Place a background view behind the library toolbar.
The titlebar is made invisible, and the placeholderView's frame is
adjusted to make room for the new view. The placeholderView's frame is
changed to fill the content view when a game is played in the library
window.

Additionally, the appearance of the faux-themed segmented control is
re-implemented to better match the appearance of the grid/list buttons.
The search field's text is now white, and the vertical positions of the
controls have been adjusted.