Move system availability toggling from OEPrefLibraryController to OEDBSystem.
This fixes an issue where systems can't be hidden in the sidebar until the preferences window is loaded. OEDBSystem methods have been added that toggle a system's enabled property if certain conditions are true, either returning an error or presenting an error in an OEHUDAlert. Additionally, the notification OEDBSystemsDidChangeNotification has been renamed to OEDBSystemAvailabilityDidChangeNotification to be more specific. The notification is now posted whenever a system's enabled property is set, regardless of whether or not the property was modified through the aforementioned toggle methods. OESidebarTogglesSystemNotification has been removed.