Increase opacity of sidebar background color view.
Currently, switching from a non-Library collection to the Library
collection causes a noticeable "snap" in color as the background of the
NSVisualEffectView goes from non-vibrant to vibrant.

To lessen that effect until a long-term solution is implemented,
increasing the background color view's opacity lessens the vibrancy as
well as the "snap." Additionally, it better matches the translucency of
other system sidebars, and it reduces the smudgy effect vibrancy can
have when colorful content is behind it--letting our console artwork
be the center of attention.