Add 'Undo Load' functionality
Whenever we Load or QuickLoad (before we load) we write to the
'OESpecialState_undo' save buffer. When the user clicks 'Undo Load',
we load from that buffer to get back where we are.

This means that when the user accidentally forgets to save for an
hour, and then loads their last save state, they can click 'Undo
Load' and get back to where they were (which may not be amazing,
but may be better than losing an hour's progress)

Note that there is only one buffer, so if you go hunting around for
a good save state and can't find one, you're still screwed.