Always move game files to trash when removing a game that was copied to the Game Library folder.
The expected behavior is that "deleted" games are actually deleted. But users tick the suppression checkboxes and forget there are two separate "delete" and "trash" dialogs which later leads to confusion. So when they "delete" a game from the Library, it still exists in the Game Library folder despite being removed from the datastore.

There is no good reason to keep "deleted" files in the Game Library folder when A) the user ~/Library folder has been hidden by default since 10.7 and B) users forget about them and end up with orphan files or, in the case of CD and M3U files, a frustrating import experience if CD files already exist in the Game Library folder.

Note: Users that have unchecked "Copy files to Game Library folder" will continue to have their files simply removed from the Library but not physically moved to trash.