Always keep games organized.
This removes the useless option to not "Keep games organized" in Library preferences which, instead of copying files to a system subfolder, would copy to an "unsorted" folder. Now all files will be automatically copied to a system subfolder when the "Copy files to Game Library folder" option is used. This should also solve the problem of files for multi-disc games sometimes getting orphaned in the "unsorted" folder during a bad import, without the need for extra logic.