Use the ROM's MD5 when naming temporary directories where to extract a game for launching.
Fixes a bug that can be reproduced with the following steps:
(1) Zipped ROM A is imported
(2) ROM A is launched
(3) ROM A is deleted
(4) Zipped ROM B is imported, and its zip file has the same filename as ROM A's zip file
(5) ROM B is launched. Expected behaviour: ROM B starts up. Actual behaviour: ROM A starts up instead.