Fix all crashes when removing unsupported plugins.
Root cause: in the very same code that tried to remove the plugins, the invocation of OECorePlugin.allPlugins caused the plugins to be loaded. Even after the plugins are removed, the loaded OEPlugin was not removed from the allPlugins list. Thus, when plugin updates were triggered, the now-deleted plugins was enumerated anyway, causing the Sparkle crash.

The plugin removal is now triggered in the plugin loading logic. This way, the plugin list cannot get out-of-sync.