Fix a crash caused by a race condition in handling AVAudioEngineConfigurationChange notifications in Catalina.
The notification callback is already active before the generator node (_gen) was instantiated, and in OpenEmuHelperApp::setupGameCoreAudioAndVideo the volume is set before the engine is actually started.
In Catalina, when the current device of the audio engine is a multi-output device (and possibly in other circumstances) this action of setting the volume of the main mixer node generates a configuration change notification, and therefore the callback is executed when _gen is still nil.
As soon as the -setDeviceAndConnection method attempts to connect _gen to the mainMixerNode, an assertion fires inside AVAudioEngine because _gen is nil, bringing down the XPC helper.

This bug was easily reproducible by creating a multiple-output virtual device in "Audio MIDI Setup.app" (even a single-device one works fine), setting it up as the default audio output device, and *then* starting a game in OpenEmu. Switching source after starting a game caused no crashes (even though switching source is partially broken at the moment)