Ensure that OEAlert modal session blocks are executed promptly.
Running NSDefaultRunLoopMode with no end date ended up preventing the execution of the blocks, unless an unrelated event was posted (like a mouse movement).

This line was originally introduced in commit dd21eabf, where it was meant to fix progress bar updates.
Since then, we started to use a stock NSProgressIndicator instead of a custom control, so it makes sense that the old fix stopped working.