When moving the game library, move all ROMs in a single operation.
Previously, at each single ROM file moved, the path in the database for that rom was changed to an absolute path.
In case of an error during the copy, this had the effect of transforming the ROMs that were copied into externally-referenced ROMs (same as you would get by disabling the "copy games into library" option in preferences). This was done implicitly and without warning, and could not be reversed.
In case of external events interrupting the process, like a power outage or an application/OS crash, the user is left a broken library becase all file references still point to inside the game library. The same breakage would happen if moving the library's Core Data storage (= database) failed.

After this commit:
(1) removing ROMs from the old game library is performed after all ROMs are copied and the database was successfully moved
(2) ROMs in the game library are never converted to absolute paths, preventing issues like #4069 from happening ever again.

Absolute paths pointing to inside the game library will be converted to relative paths to retroactively patch out damage from issue #4069