fix: Remove NSCursor hide / unhide logic with full screen transitions
The behaviour remains the same, as entering and exiting full screen
still hide the `NSCursor`, due to a call to
`[OEGameControlsBar hideAnimated:]`, which in turn calls
`[NSCursor setHiddenUntilMouseMoves:YES]`.

The bug manifests on macOS 11.0.1 as the user being unable to restore
cursor visibility. It appears to be an issue due to calls to `NSCursor`
`hide` and `unhide`, interleaved with calls to
`[OEGameControlsBar hideAnimated:]`. The only way to restore the cursor
is to enter Spaces or some other trickery.