Automatically map controls for Xbox One S Wireless Controller (Model 1708) with BLE firmware.
Since firmware v5.11, support for Bluetooth Low Energy was added which changed device Product ID to 0xB20 and the "View" and "Xbox" button HID usage elements, thus requiring a new automapping profile. This HID profile is identical to that of the Model 1914 -- the Model 1708 device only lacks the new "Share" button of the Model 1914 device.