Add 8BitDo GBros. Adapter to controller database
Assumes that the connected controller in X(Input) mode is a GC controller, as there is no way of knowing what kind of controller is connected and the supported classic controllers are already automapped perfectly in S(witch) mode. They still work in X(Input) mode, but the button labels won’t match.